src/is-object-detector/pyproject.toml
[tool.poetry]
name = "is-object-detector"
version = "0.1.0"
description = ""
authors = ["Luiz Carlos Cosmi Filho <luizcarloscosmifilho@gmail.com>"]
readme = "README.md"
packages = [{include = "is_object_detector"}]

[tool.poetry.dependencies]
python = "^3.8"
pyturbojpeg = {git = "https://github.com/lilohuang/PyTurboJPEG.git"}
numpy = "1.22.4"
opencv-contrib-python = "4.7.0.68"
vine = "1.3.0"
is-wire = {git = "https://github.com/labviros/is-wire-py.git", rev = "add_communication_tracing"}
is-msgs = "1.1.10"
python-dateutil = "2.8.0"
opencensus-ext-zipkin = "0.2.1"
onnxruntime = "^1.15.1"

[tool.poetry.scripts]
is-object-detector = 'is_object_detector.service:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
